# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScreenCrew API - friend-group screen sharing app
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
        "403":
          description: Registration not permitted (closed, server full, or invalid invite)
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with existing credentials
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthInput"
      responses:
        "200":
          description: Logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /admin/claim:
    post:
      operationId: claimAdmin
      tags: [admin]
      summary: Become admin by presenting the server's configured admin password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimAdminInput"
      responses:
        "200":
          description: Admin granted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
        "403":
          description: Admin password not configured or incorrect
...
  /admin/config:
    get:
      operationId: getServerConfig
      tags: [admin]
      summary: Get server configuration (admin only)
      responses:
        "200":
          description: Server configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerConfigInfo"
        "403":
          description: Forbidden
  /admin/invites:
    get:
      operationId: listInvites
      tags: [admin]
      summary: List server invite keys (admin only)
      responses:
        "200":
          description: Invite keys
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ServerInvite"
        "403":
          description: Forbidden
    post:
      operationId: createInvite
      tags: [admin]
      summary: Create a server invite key (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteInput"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServerInvite"
        "403":
          description: Forbidden
  /admin/invites/{id}:
    delete:
      operationId: revokeInvite
      tags: [admin]
      summary: Revoke a server invite key (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked
        "403":
          description: Forbidden
        "404":
          description: Not found
...
  /rooms/{roomId}/channels:
    get:
      operationId: getChannels
      tags: [channels]
      summary: List channels in a room (private channels only for staff)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      operationId: createChannel
      tags: [channels]
      summary: Create a channel (owner/mod only)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateChannelInput"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "403":
          description: Insufficient permissions
  /rooms/{roomId}/channels/{channelId}:
    patch:
      operationId: updateChannel
      tags: [channels]
      summary: Update a channel (owner/mod only)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelInput"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "403":
          description: Insufficient permissions
        "404":
          description: Channel not found
    delete:
      operationId: deleteChannel
      tags: [channels]
      summary: Delete a channel (owner/mod only; cannot delete last channel)
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: integer
        - name: channelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Insufficient permissions
        "404":
          description: Channel not found
        "409":
          description: Cannot delete the last channel
...
components:
  schemas:
...
    AuthInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 2
        password:
          type: string
          minLength: 4
    RegisterInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          minLength: 2
        password:
          type: string
          minLength: 4
        inviteKey:
          type: string
          nullable: true
          description: Required when the server registration mode is "invite".
...
    ServerConfigInfo:
      type: object
      required: [serverName, registration, maxUsers, userCount]
      properties:
        serverName:
          type: string
        registration:
          type: string
          enum: [open, invite, closed]
        maxUsers:
          type: integer
        userCount:
          type: integer
    ServerInvite:
      type: object
      required: [id, key, createdBy, uses, createdAt]
      properties:
        id:
          type: integer
        key:
          type: string
        createdBy:
          type: integer
        maxUses:
          type: integer
          nullable: true
        uses:
          type: integer
        expiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    CreateInviteInput:
      type: object
      properties:
        maxUses:
          type: integer
          minimum: 1
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
    ClaimAdminInput:
      type: object
      required: [adminPassword]
      properties:
        adminPassword:
          type: string
          minLength: 1
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, username, isAdmin, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        displayName:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        steamUrl:
          type: string
          nullable: true
        discordUrl:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        nameColor:
          type: string
          nullable: true
        avatarStyle:
          type: string
          nullable: true
        isAdmin:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    Channel:
      type: object
      required: [id, roomId, name, type, position, isPrivate, createdAt]
      properties:
        id:
          type: integer
        roomId:
          type: integer
        name:
          type: string
        type:
          type: string
          enum: [text, voice, announcement, media]
        position:
          type: integer
        isPrivate:
          type: boolean
        minViewRole:
          type: string
          enum: [owner, mod, member]
        minSendRole:
          type: string
          enum: [owner, mod, member]
        createdAt:
          type: string
          format: date-time
    CreateChannelInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 40
        type:
          type: string
          enum: [text, voice, announcement, media]
        isPrivate:
          type: boolean
        minViewRole:
          type: string
          enum: [owner, mod, member]
        minSendRole:
          type: string
          enum: [owner, mod, member]
    UpdateChannelInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 40
        type:
          type: string
          enum: [text, voice, announcement, media]
        isPrivate:
          type: boolean
        position:
          type: integer
        minViewRole:
          type: string
          enum: [owner, mod, member]
        minSendRole:
          type: string
          enum: [owner, mod, member]
...